HH-101495 rename package according to classes' names
