+++ nab-tests/src/test/java/ru/hh/nab/kafka/KafkaTestConfig.java
@@ -5,9 +5,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import ru.hh.nab.common.properties.FileSettings;
-import ru.hh.nab.kafka.consumer.DefaultListenerFactory;
-import ru.hh.nab.kafka.consumer.DeserializerSupplier;
-import ru.hh.nab.kafka.consumer.ListenerFactory;
+import ru.hh.nab.kafka.listener.DefaultListenerFactory;
+import ru.hh.nab.kafka.listener.DeserializerSupplier;
+import ru.hh.nab.kafka.listener.ListenerFactory;
 import ru.hh.nab.kafka.publisher.PublisherFactory;
 import ru.hh.nab.kafka.publisher.SerializerSupplier;
 import ru.hh.nab.kafka.serialization.JacksonDeserializerSupplier;
